feat(host): Mount ISO image for kickstart installation
The kickstart process requires extracting kernel files (e.g., `initrd`,
`vmlinuz`) from the ISO to customize QEMU boot parameters. This change
mounts the ISO to a temporary directory to enable this.

Cleanup of the temporary mount directory is also needed, but is not
implemented yet and is planned for a future update.